improving smart contracts

+++ src/interface/IKOL.sol
@@ -7,9 +7,9 @@ interface IKOL {
 
     function disableKOL(uint256 _idKOL) external;
 
-    function getKOLCampaign(uint256 _id) external returns (address,uint256);
+    function getKOLCampaign(uint256 _id) external returns (address, uint256);
 
     function enableKOL(uint256 _idKOL) external;
 
-    function isValidKolCampaign(uint _id) external returns(bool isValid);
+    function isValidKolCampaign(uint256 _id) external returns (bool isValid);
 }
